Beginning of an implementation of a dialog for setting arbitrary breakpoints.
Currently it has a text entry for 'filename' and one for 'line number' and
attempts to set a breakpoint at the requested line.  There's not really any
error handling yet, but it works for simple cases.  Eventually there should be a
way to set a breakpoint on a function name as well -- not sure whether that
should be a separate option in the UI or combined somehow.

svn path=/trunk/; revision=440