Use the GNOME system monospace font instead of a generic 'monospace' font.
* src/persp/dbgperspective/glade/preferencesdialog.glade: minor re-wording
* src/persp/dbgperspective/nmv-dbg-perspective.cc: Get the system monospace
font from gconf and use that instead of setting it to a generic 'monospace' font
name.
* src/workbench/nmv-workbench.cc: In order to be able to update our font
when the system monospace font changes, I had to also watch the gconf directory
/desktop/gnome/interface for configuration changes.  This seems a bit ugly, but
IConfMgr doesn't currently have any other way to do this.  Maybe we need to add
something to IConfMgr to accomodate this, but for now I just did the
quick-and-dirty fix.

svn path=/trunk/; revision=478