Disable 'toggle enable breakpoint' when there's no breakpoint on the line
Add handling for when the cursor position changes so that we can enable or
disable certain actions depending on what line we're on.  Now, the "toggle
enable breakpoint" action is set insensitive when there's no breakpoint set on a
particular line.  This is a partial fix for bug #430346.  The method I used to
implement this feels a little bit 'brute-force'-ish, so if you know of better
ways to accomplish this, let me know.

svn path=/trunk/; revision=484