first tentative to fix 534695
Nemiver actually is slow when attaching to some targets.
This comes from the fact that nemiver steps over the first C instruction it
finds when attaching to a program. Stepping over might take a lot of time
if for instance there is no debug information for the symbol over which we are trying
to step over.

In this fix tentative, nemiver instead steps into the next machine
instruction instead of the stepping over the first C instruction

svn path=/trunk/; revision=821