Support graphically adding a variable to the monitor
[This is not a proper commit entry.  Commits on this branch are
all going to be re-hashed anyway.]

The variable inspector dialog now has a button that lets you add
the current variable/expression to the monitor.  It's very rough as no
check is performed to make sure you can't add the same expression to the
monitor twice, etc.  Also, once the variable is added, we face an issue
that is explained in #663959, namely the life time management of expressions
backed by variable objects.  I need to address that issue separately
before getting back to this business.