Refer to the 'tests" sub-directory explicitly
In the test-libtool-wrapper-detection.cc, we assume that the tests are
run from within the "tests" sub-directory of the build directory.
That is, the current directory should be $builddir/tests when the
tests are run.

This is wrong because invoking that test from anywhere should work by
doing:

    make -C <builddir> check TESTS=runtestlibtoolwrapperdetection

So to this this, I am changing the code of the tests to provide an
absolute path to the "tests" directory.

	* tests/test-libtool-wrapper-detection.cc (test0)
	(test_filename_with_dashes): Refer to the "tests" sub-directory
	(of either the source directory or the build directory)
	explicitly.

Signed-off-by: Dodji Seketeli <dodji@seketeli.org>